Move sub-component-local UI state out of UsageFilterPanel
Member group selection, model complexity dropdown state, and
selection-summary collapse state only matter to their own
sub-component, so own them there instead of threading props down
from the parent.
